Fix #267: Extract Application class into focused classes
