+++ spec/spec_helper.rb
@@ -9,9 +9,21 @@ require 'webmock/rspec'
 # Load the application
 require_relative '../lib/ghb'
 require_relative '../lib/ghb/application'
+require_relative '../lib/ghb/aws_job_builder'
+require_relative '../lib/ghb/code_deploy_job_builder'
+require_relative '../lib/ghb/dependabot_manager'
+require_relative '../lib/ghb/dockerhub_manager'
+require_relative '../lib/ghb/file_scanner'
 require_relative '../lib/ghb/github_api_client'
+require_relative '../lib/ghb/gitignore_manager'
+require_relative '../lib/ghb/language_job_builder'
+require_relative '../lib/ghb/licenses_job_builder'
+require_relative '../lib/ghb/linter_job_builder'
 require_relative '../lib/ghb/options'
+require_relative '../lib/ghb/repository_configurator'
+require_relative '../lib/ghb/slack_job_builder'
 require_relative '../lib/ghb/status'
+require_relative '../lib/ghb/variables_job_builder'
 require_relative '../lib/ghb/workflow/job'
 require_relative '../lib/ghb/workflow/step'
 require_relative '../lib/ghb/workflow/workflow'
